requirements.txt
fastapi
uvicorn
passlib
bcrypt
python-jose
peewee
python-multipart
pyjwt